gooftool: fix EC/BIOS wpstatus in report
In http://codereview.chromium.org/6778007 the execution order of
write-protection status probing is changed to prevent EC issue,
however the text message is not changed.

This CL corrects the output message, and isolate the execution commands so that
we may disable checking of EC on ARM platforms in future.

BUG=chrome-os-partner:854
TEST=gooftool --finalize --db_path DBPATH --upload_method none
     flashrom -V | grep Found
     # the output is SPI(BIOS), not LPC(EC).
     reboot # success on Z*

Change-Id: Iebe80a84a36349d7dcac006fb6cbcba3932ce069

R=yjlou@chromium.org

Review URL: http://codereview.chromium.org/6793006