factory_test_tools: support legacy probing mode
The new heuristic probing for USB/PCI devices is good, but it is possible to
have some device that cannot be probed by the sysfs rules.

To solve this issue, a new "legacy mode" is introduced: if an approved
component list is simply full of traditional USB/PCI device info, we should
ignore our heuristic search results and simply compare if the one of the
component in list can be found on system.

ex: id_wireless: ['1a1b:2b3c' '0030:1234],
    and the system has 0030:1234 installed but not as 'wlan0'.

Before this fix, probing will result in id_wireless: [''] and give failure,
after this fix probing will give ['0030:1234'] and success.

BUG=chromium-os:13967
TEST=gft_hwcomp.py MODIFIED_COMPONENT # to have legacy fields only

Change-Id: Ib3fb8d5a01bfb24fd84c10d611d9ff6f8637ef23

R=waihong@chromium.org

Review URL: http://codereview.chromium.org/6821072