Add error reporting to data parsing and verification.
Addressed two TODOs.  This allows for more human-friendly
modification of data schemas and data, because now it will
tell you where and what kind of errors are present in the
data.

Also fix up the order of schema elements to match the
order in yaml data.

BUG=chrome-os-partner:6003
TEST=run hwid_tool overview or data verification

Change-Id: I838831d61c7777d1eaa74d4457ddb0908276469f
Reviewed-on: https://gerrit.chromium.org/gerrit/17308
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>