Allow generation and input of machine-readable component data.
Add hwid_tool command options to produce probe data
in vanilla yaml format, which can then be fed back into
another command which will add that to the component
database.  The hwid_tool behavior is modified to always
update the database files with data changes.  This is
as opposed to creating a series of copies with incremental
changes (which is not very useful, since we have git history).

BUG=chrome-os-partner:6003
TEST=hwid_tool w/ assimilate_probe_data, probe_device, or
     board_create

Change-Id: I2bc4f4e20d8f2277dcb0fad57f3aef81daa1f75c
Reviewed-on: https://gerrit.chromium.org/gerrit/17392
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>