Clean up status data repr, remove crufty fields, and /s/hash/volatile.
Status is easier to deal with (and extend) if consolidated into
a single dict for hwids and one for components, respectively.

The bitmap path and release string fields are removed.  Bitmaps
are obsolete.  For release, we can represent the same data within
either initial_config or volatile.

Volatile is not restricted to just containing hash values.  Change
the name to reflect that (which also makes fields sort better).

BUG=chrome-os-partner:6003
TEST=run hwid_tool usng data files that contain fields matching
     the changes in this CL (eg, CompDb with 'registry' and
     'status_map' as opposed to the more numerous previous fields.
     successfully running any hwid_tool command is sufficient
     to test, since it will validate data on any read.

Change-Id: I7b349781093358cce49089feff147705bdd17206
Reviewed-on: https://gerrit.chromium.org/gerrit/18446
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>