Create indpendent FW hash generation function.
Pull out the hash generation code from probe() and allow
it to be called independently.  This new function can be
run on arbitrary firmware blobs and it will try to determine
the type of fw (main vs ec) and produce the correct corresponding
hashes.

BUG=chrome-os-partner:8616
TEST=import the probe module and run probe() with arguments ;
     alternatively, wait for the follow-on CL which includes
     the necessary arguments in the gooftool probe command.

Change-Id: Idc43da9a709aa5274bd19377a144a668a5cdada8
Reviewed-on: https://gerrit.chromium.org/gerrit/19212
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>