Update gooftool to use the new event_log mechanism.
Since the event log code lives in the autotest repo (for now),
some import magic was needed, which should be removed when
that code moves here.

Gooftool now creates an event log on each execution and uploads
a tarball of the logs, plus the factory log, to replace the
previous report uploading process.

BUG=chrome-os-partner:9965
TEST=run './gooftool upload_report' on a DUT

Change-Id: If92bb31dd39b72f77619846beb3c36d509078538
Reviewed-on: https://gerrit.chromium.org/gerrit/23619
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>