hwid_tool: cooked probe results class, lists in probe results
Probes not return lists of independent components if more than one
component is present for a component class; this is in contrast to
previous behavior where all present components were mashed into a
single string probe result, which then needed some kind of combined
name.

Update assimilate command to match the new datastructures.

Remove crufty commands that were not used (create_hwids and associated
cross product calculation).  This functionality is easy to re-create
with the new datastructures, but is currently unused.

Rename xxxData database classes into xxxSpec, as this is more reflective
of their role.

[This CL is part of a larger refactoring series; functionality
is known to be broken unless the whole series is applied]

BUG=chrome-os-partner:10319
TEST=hwid_tool is probably broken, apply later commits before testing

Change-Id: Ie414ac2c3463462a0ccf4a9e5e41c58eb191faba
Reviewed-on: https://gerrit.chromium.org/gerrit/26422
Commit-Ready: Tammo Spalink <tammo@chromium.org>
Reviewed-by: Tammo Spalink <tammo@chromium.org>
Tested-by: Tammo Spalink <tammo@chromium.org>