factory: Fix connection manager and goofy bugs
The CL fixes two bugs. The first one is that goofy does not construct
the connection manager with the wireless LAN setting in the test list.
The second one is that we need to stop/start other related services such
as wpasupplicant and flimflam_respawn in the connection manager to get
a clean test environment.

BUG=chrome-os-partner:11735
TEST=Tested on real devices -> works.

Change-Id: Ic831f2aca2f21911703214eaf544f859c102039c
Reviewed-on: https://gerrit.chromium.org/gerrit/28256
Reviewed-by: Jon Salz <jsalz@chromium.org>
Tested-by: Tai-Hsu Lin <sheckylin@chromium.org>
Commit-Ready: Tai-Hsu Lin <sheckylin@chromium.org>