hwid_tool: Unmatched components should fail to match bom
When we match bom using probed components, we should not only check the
extant components do match. There might be some components not listed
in component_db, and are not matched to what the bom specified.
This CL fix MatchComponentSpecWithData such that it will check all the
components of this bom can find a matched probed component.

Change-Id: I3848230283b40ca5c5483d20bcae43b471779275

BUG=chrome-os-partner:14176
TEST=plug a wrong wifi not listed in component_db
     check gooftool verify_hwid will fail.

Change-Id: I54bfe18c5c3923e6e38f817a72b838e78a73b399
Reviewed-on: https://gerrit.chromium.org/gerrit/33509
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>