ExtDisplay: Fix bug of detecting/switching external display.
The original method of detecting external display is buggy:
If main display and external display's IDs have the same postfix, it
cannot tell that if external display is plugged-in or not.

Another bug is that in some board, if we turn on external display and
turn off main display simutaneously, it fails. The workaround is to
turn on dual displays first then turn off the main one.

BUG=None
TEST=manual
manually tested on Stout and Link.

Change-Id: I2da90dfe1a647684f57562283630e7875db2a993
Reviewed-on: https://gerrit.chromium.org/gerrit/38725
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Ready: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>