Run unittests in parallel to shorten "make test" time.
Right now we have 30+ unittests, it is time to make them to run in
parallel. To deal with tests which cannot run in parallel, we provide
an argument "--isolated" to specify tests which need to run
sequencially. Also, it provides a fallback mechanism that re-run failed
test sequencially.

Now with CROS_FACTORY_ROOT set to a temp directory for each tests, and
a fix for shopfloor_standalone_unittest, we can run all unittests in
parallel now.

BUG=chrome-os-partner:16464
TEST=make test

Change-Id: Idaadf7e478c6b9aa6b2a100aff9d6ea8bc9955e1
Reviewed-on: https://gerrit.chromium.org/gerrit/38987
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Ready: Shuo-Peng Liao <deanliao@chromium.org>