FactoryTest: add disable_abort to disable 'abort test ...' item from menu
For some test we do not want user to abort the test or its subtest from menu.
With disable_abort as an argument of FactoryTest, now we can control the
appearance of menu item "Abort active test xxx and continue testing".
If goofy is in engineering mode, there is no restriction to abort the
test.

BUG=chrome-os-partner:17346
TEST=run test with argument disable_abort = True. While test is running,
there is no item "Abort active test xxx and continue testing" in the
menu when user click on the test or its parent node.

Change-Id: Ifbf85a5ebdbebf1f7e040860c20bf86c59953260
Reviewed-on: https://gerrit.chromium.org/gerrit/41923
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>