summary.py: add an darg disable_input_on_fail
Use disable_input_on_fail to disable user pass/fail when the overall
status is not PASSED. This argument is useful when we want to use
summary test as a barrier.

BUG=chrome-os-partner:17346
TEST=run summary test with darg disable_input_at_failure. When the
overall status is not PASSED, user can not pass/fail the test.
When the overall status is PASSED, user can pass the test.

Change-Id: Iee8d08668fb7529a131661d5a1069414e36871da
Reviewed-on: https://gerrit.chromium.org/gerrit/41925
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>