Factory: Add manual override option to HWID test.
This allows an operator to change the HWID returned by the shopfloor
server. Turned off by default. The main use of this option is for
RMA. The shopfloor server returns the original HWID of the
machine which an operator may adjust if they have changed components.

BUG=chrome-os-partner:17302
TEST=Set 'manual_override: True' in a test_list and use a
shopfloor server to provide a HWID. Verify you get prompted to select
a HWID with the default option being the HWID from the shopfloor
server.

Change-Id: Ie7655af3001655fa21200c63610fabd0f66c15d3
Reviewed-on: https://gerrit.chromium.org/gerrit/42343
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Dave Parker <dparker@chromium.org>
Tested-by: Dave Parker <dparker@chromium.org>