wireless.py: Basic connectivity test for Wifi and LTE.
BUG=chrome-os-partner:17122
TEST=Manual

In testing in Mountain View, when this LTE test connected with Verizon
and then attempted to connect to http://www.google.com, it would
reliably fail to open the url. This is after having successfully
connected to Verizon Wireless. A delay was needed before trying to
send data traffic. Rather than using a hard coded delay I changed
the previous 10 second timeout in urlopen() to 2 seconds and put it
into a 5 reps loop. Also added an exception handler for URLError.

Change-Id: Ib2be14dc0fcf35ee01383984806499ca821aee2a
Reviewed-on: https://gerrit.chromium.org/gerrit/43248
Reviewed-by: Dave Parker <dparker@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Commit-Queue: Walter Murphy <wmurphy@google.com>
Tested-by: Walter Murphy <wmurphy@google.com>