Implement a timeout only mode for the camera test
In some cases we just want the camera to run for a period of time without
any user input for pass/fail, such as for run in testing. This will allow
a test list that just specifies an amount of time for the camera to run.

Note this can still fail if the camera does not initialize properly.

BUG=chrome-os-partner:17634
TEST=Manual, run the test list below on a DUT with a camera.

OperatorTest(
    label_en='Camera RunIn',
    label_zh=u'相机',
    pytest_name='camera',
    dargs={'face_recognition': False,
           'timeout_secs': 10,
           'timeout_run': True}),

Change-Id: Ida7bfbcca751127ded7a75e3c37f3adf7abe94bd
Reviewed-on: https://gerrit.chromium.org/gerrit/43762
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>