add path_loss computation in radiated_cellular_gobi
Compute the path_loss when category is 'calibration'. A path_loss is
only defined when the expected power is known in prior.
Expected power have to be defined in separate parameter,
calibration_target. A GetUniqueIdentification() is added to identify the
module so to map the corresponding target.

This change also factors out some common code and moves the
DownloadParameters from radiated_cellular_gobi to rf_framework.

BUG=chrome-os-partner:17967
TEST=Setup a fake calibration_target and manually on DUT.

Change-Id: I6ae424842df520d644a3aaef40d5e096f6d54130
Reviewed-on: https://gerrit.chromium.org/gerrit/45134
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Chun-Ta Lin <itspeter@google.com>
Tested-by: Chun-Ta Lin <itspeter@google.com>