depthcharge: Add a script to update netboot settings.
This script updates settings stored in the SHARED_DATA section of the image for
retrieval by depthcharge when netbooting. The format of the params is
documented in a commit message in depthcharge.

BUG=chrome-os-partner:16688
TEST=Used this script to update the netboot parameters of a depthcharge image
and saw that they were recognized and were correct. Saw that existing
parameters were recognized and could only be overwritten when --clobber was
set. Verified the behaviors of the --input and --output options.

Change-Id: I262e3cc634e19f60a092d487b767e9312f42afac
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/45946
Reviewed-by: Vic Yang <victoryang@chromium.org>
Tested-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>