For SerialEchoTest, allow user to specify (send, recv) command tuple.
It is used to let DUT send a command to a fixture and expect its response.
With customized (send, recv), we can use it to issue command to the fixture.

This CL also handles serial read/write timeout exception.

BUG=chrome-os-partner:18116
TEST=make test
Also tested on DUT with a fixture which echos command it receives.

Change-Id: Iad57246c48e162f424edc2115233e05644590e49
Reviewed-on: https://gerrit.chromium.org/gerrit/45990
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>