Refactor the device argument and add automated device probing
The device argument can be more properly defined as the path to a given
device instead of a device name, this enables it to be used with the file
mode as a direct path to the file.

As any device that has a /dev/mmcblk0 device without any SD card inserted
uses eMMC we can probe for that automatically.

BUG=chrome-os-partner:18468
TEST=Manual, run test on DUT.

Change-Id: I3f36cd030c79335dbf9198a61983c841590206b4
Reviewed-on: https://gerrit.chromium.org/gerrit/47708
Reviewed-by: Dave Parker <dparker@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>