Minijack: Add TestParser and helper methods for manipulating database
This change adds the TestParser. As its logic is quite simple, we use
it as an example to demonstrate the helper methods for manipulating
the database.

This set of helper methods hide the underlying SQL commands. A parser
just needs to pass a Python dict to them.

BUG=chrome-os-partner:18691
TEST=Manual
Run Minijack and check the old log converts into the Test table.

Change-Id: Ic9551e51f7c677d859375db0c269e41343c25fb5
Reviewed-on: https://gerrit.chromium.org/gerrit/48097
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>