Minijack: Add Event and Attr parsers
An event has a tree structure which contains multi-level dicts or lists.
We need to flatten them and record them in the Attr table.

So an event is put to one row in the Event table and multiple rows in the
Attr table.

BUG=chrome-os-partner:18691
TEST=Manual
Run Minijack and check the old log converts into the Event and Attr tables.
Run 'python -m doctest -v parser_base.py' passed.

Change-Id: I150d7df9e3310a92ef6ddb6270ecbddea17f52dd
Reviewed-on: https://gerrit.chromium.org/gerrit/48098
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>