Enhance suspend resume test for late suspend reliability
In some cases the suspend_resume test can effectively hang forever in suspend
if the actual suspend is delayed for some reason past the wake alarm time.
This seems to most commonly happen when some other part of the system crashes
and delays suspend while generating crash dumps and such, but it could be
caused by other loading issues at suspend time.

This change allows the test to continue to push out the wake alarm if no
suspend is detected, and adds a new parameter for a worst case suspend time,
defaulting to 60s.

BUG=chrome-os-partner:18518
TEST=Manual, run test on DUT.

Change-Id: I5db6f4d4a236a9050494e9928c36f9021de4110b
Reviewed-on: https://gerrit.chromium.org/gerrit/48500
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>