Minijack: Abstract the data model of Minijack
The first version of Minijack defined the table schema in a Python dict,
for example, the Attr table:
    schema_dict = {
      'device_id': 'TEXT',
      'time': 'TEXT',
      'attr': 'TEXT',
      'value': 'BLOB',
    }
    self.setup_table('Attr', schema_dict,
                     primary_key=['device_id', 'time', 'attr'])

This design is rough and tends to buggy if a developer types a wrong name.
This chain of CLs tries to improve the Minijack data model by introducing
some well-defined field and model classes, which helps validating the
data types during assignments and also hides the underlying database
implementation. The new database usage will become:

    class Attr(db.Model):
      device_id = db.TextField(primary_key=True)
      time = db.TextField(primary_key=True)
      attr = db.TextField(primary_key=True)
      value = db.BlobField()

    table = database.GetOrCreateTable(Attr)
    row = Attr(device='xxx', time='xxx', attr='xxx', value='xxx')
    table.InsertRow(row)

This CL introduces the Field and Model classes. Check the detailed usages
in the unit tests.

BUG=chrome-os-partner:18749,chrome-os-partner:18798
TEST=make test

Change-Id: I17650b468b038dc7d3f2989048296245cc780aea
Reviewed-on: https://gerrit.chromium.org/gerrit/48884
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>