Minijack: Just insert the row to the Event and Attr tables
The original logic is to check the row exists first to determine if it
is an insertion or an update. However, Event and Attr processing
happens on every event. It takes time to do this check.

So this CL just inserts the row. If this row already exists in the table
(technically, the primary key already exists), which is caused by
duplicated log, it catches the exception and ignores it.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 10m23s
        new: 9m52s
       gain: 5%

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: I8351dd08ef797a11d7bc4b48ae18131e8265a07a
Reviewed-on: https://gerrit.chromium.org/gerrit/49023
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>