Minijack: Use a list of path instead of a string
The native string concatenation during the recursive FlattenAttr() method
is expensive. This CL changes the output of FlattenAttr to a list of
strings instead of one concatenated string. It then does a join() to the
list to produce the original string.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 57.47 ~ 57.90s
        new: 56.86 ~ 57.45s

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: I530d4fc5e2f5f4f43821ed881b4ccdfee6eaf3d2
Reviewed-on: https://gerrit.chromium.org/gerrit/49037
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>