Minijack: Utilize the sqlite3 executemany() API
One event contains in average ~30 flattened attributes. So the original
execute() API is not quite efficient. This CL adds the executemany() API
support. So the insertion time is reduced dramatically.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 10m23s
        new: 58s
       gain: 91%

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: I096b212e470fe891749a23cc69ab373f5b682105
Reviewed-on: https://gerrit.chromium.org/gerrit/49036
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>