Minijack: Use MEMORY journaling mode which saves disk I/O
For speed-up purpose, the journal is actually not necessary in Minijack.
It does not rollback a transaction.

Quote from Sqlite document:
The MEMORY journaling mode stores the rollback journal in volatile RAM. This
saves disk I/O but at the expense of database safety and integrity. If the
application using SQLite crashes in the middle of a transaction when the
MEMORY journaling mode is set, then the database file will very likely go
corrupt.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 57s
        new: 29s
       gain: 50%

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: I3288cdb5af9bdff26fed3e18b1376382512c7b32
Reviewed-on: https://gerrit.chromium.org/gerrit/49189
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>