Minijack: Don't wait OS to write all content to disk before the next action
The data of Minijack is not quite critical. We can easily reconstruct it from
the raw logs. So we skip this sync for speed-up.

Quote from Sqlite document:
With synchronous OFF (0), SQLite continues without syncing as soon as it has
handed data off to the operating system. If the application running SQLite
crashes, the data will be safe, but the database might become corrupted if
the operating system crashes or the computer loses power before that data
has been written to the disk surface.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 29s
        new: 16s
       gain: 45%

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: I747cc687cb11dc62158f3af84595418c95910d33
Reviewed-on: https://gerrit.chromium.org/gerrit/49190
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>