Minijack: Do duplication check only using Event; remove primary key for Attr
Removing primary key on Attr table reduces insertion time as there is no
duplicaton key. But we still want to remove the duplicated rows. As the
rows in Event and Attr come from the same source. So we only do the
duplication check on Event.

It merges EventParser and AttParser into one EventAttrParser.

The result of parsing data and dumping it to database:
  data size: 2MB
   original: 15.7s
        new: 12.8s
       gain: 19%

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: Ifdf15760c0e4fcdb1987913c31eb988a33c56785
Reviewed-on: https://gerrit.chromium.org/gerrit/49462
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>