Minijack: Refine the name of parsers to exporters
Originally, we call the class, which extracts the knowledge from the event
and dumps to the database, a parser. This name "parser" is quite general.
It may be confused with the YAML-to-Python parser. The name "exporter" is
more accurate to describe its behavior.

BUG=chrome-os-partner:19047
TEST=Manaul
Run Minijack and check its result the same as before.

Change-Id: I0ae3b504baa64322376ce2658f0c2ade31e2bfcc
Reviewed-on: https://gerrit.chromium.org/gerrit/49581
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>