Minijack: Introduce EventPacket which is passed to the exporters
An EventPacket is a non-preamble event combined with its preamble. It is
used as an argument to pass to the exporters.

FlattenAttr() and FindAttrContainingKey() are now moved to this class.

Unit tests are added.

BUG=chrome-os-partner:19047,chrome-os-partner:18798
TEST='make test' and manaul
Run Minijack and check its result the same as before.

Change-Id: I1614b1e1d19c4b0401ebcf4dfe52b4015d2640c7
Reviewed-on: https://gerrit.chromium.org/gerrit/49584
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>