Minijack: Refactor EventReceivingWorker and remove Cleanup methods of exporters
For preparing multi-processing Minijack, this CL refactors the event receving
code to the EventReceivingWorker class, which is callable, to receive an event
stream and dump to database.

Also removes the Cleanup methods of exporters, as no one uses it.

BUG=chrome-os-partner:19047
TEST=Manual
Run Minijack and check its result the same as before.

Change-Id: I8b1d52e6515f01e0b28326142396a9c7c2630ae7
Reviewed-on: https://gerrit.chromium.org/gerrit/49845
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>