Minijack: Speed-up by only reading the first event to find the preamble
On shopfloor service, the event blob is appended to an existing file
incrementally. So if no preamble in an event blob, we should find it from
the log file. Actually, we only need to read the first event to retrieve
the preamble, instead of reading the whole file.

The result of processing multiple-day logs, offline.
  data size: 410MB
   original: 8m12s
        new: 6m44f
       gain: 18%

The online result is better as the log is coming incrementally.

BUG=chrome-os-partner:18799
TEST=Manual
Checked Minijack still works as before.

Change-Id: If8bc7e824376f8e50243fe448d1de194d01242ef
Reviewed-on: https://gerrit.chromium.org/gerrit/49869
Reviewed-by: Andy Cheng <andycheng@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>