Minijack: Check the table exists before creating it
Don't rely on the 'CREATE TABLE IF NOT EXISTS' SQL command. Do the checking
in the Database class such that we can have more controls, like verifying
the table schema, which will be done in the later CL.

This CL also introduces the sqlite_master table which keeps all the info of
the tables in the Sqlite database.

BUG=chrome-os-partner:18930
TEST=make test

Change-Id: I2050ff94328c9c080f966a27ac7c34058c07aa09
Reviewed-on: https://gerrit.chromium.org/gerrit/50152
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>