Minijack: Automatically get the proper table from database
Minijack assumes that one model is a table. So the model type tells what
the table this model belongs to.

To simplify our database APIs, we now hide the Table object. Any access
to the Table is automatically done by the Database APIs. In the other
words, the Database APIs now gets the proper table and does the jobs on
that table. The programming logic can be simplified from:
  new_event = Event(...)
  table = database.GetOrCreateTable(new_event)
  table.InsertRow(new_event)
to:
  new_event = Event(...)
  database.Insert(new_event)

BUG=chrome-os-partner:19194
TEST=make test
TEST=Manual
Run Minijack on the build log and check it produces the same result.

Change-Id: Id53f46ed86a00940423112dbc11edd7551d46465
Reviewed-on: https://gerrit.chromium.org/gerrit/50752
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>