Retry logic for getting RSSI in cellular_gobi_rssi test
The modem might not be ready at the time we ask for a RSSI. This CL
try to address this issue by adding a retry logic for specific modem
response. In addition, the net_utils.PollForCondition function is
generalized for returning values to the caller so caller can reuse
important results without re-executing condition function.

BUG=chrome-os-partner:19271
TEST=Manually run it on DUT several times and verified by seeing
restricted FTM.

Change-Id: I83404d6696328e2964180fe2aba101c480296ad1
Reviewed-on: https://gerrit.chromium.org/gerrit/50612
Commit-Queue: Chun-Ta Lin <itspeter@google.com>
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Tested-by: Chun-Ta Lin <itspeter@google.com>