Minijack: Implement Minijack Archiver
Minijack is a long live service to process all the log. The total volume of
devices is a lot. If we keep all the log there, the database size is too large
that any insertion/query becomes very slow. We should archive the log of the
finalized devices into other database, on a per-day or per-week basic.

Minijack Archiver is a class to help doing the above jobs easily. It checks
the Device table to find all devices finalized before a given date and not
archived even. And dumps all Event/Attr records to a backup database.

Also introduces a new minijack_status field in the Device model, which keeps
the processing status of Minijack. So far, it is either '' or 'ARCHIVED'.

BUG=chrome-os-partner:19194
TEST=make test

Change-Id: I3c6f4c83d56ad9ce104b49a3780d4200f9507fd3
Reviewed-on: https://gerrit.chromium.org/gerrit/50753
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>