Minijack: Generate a new field event_id as the unique primary key of Event
The original primary key of Event is the fields, device_id and time. As the
time value is not reliable to ensure the order when syncing time happens.

So we need to generate a new field event_id as the unique primary key of
Event. We use the log filename + SEQ as the event_id.

BUG=chrome-os-partner:19333
TEST=make test
TEST=Manaul
Run Minijack on the build log and check it produces the same result.

Change-Id: I6f8dc52dd7a4e2b98cf25e06bf24921e25d9d901
Reviewed-on: https://gerrit.chromium.org/gerrit/50773
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>