Minijack: Keep the event_id on Test table, which helps refering Event easily
As we don't use the device_id + time as primary key on Event, so the original
start_time and end_time are no longer able to refer Event. We keep the
event_id on Test table, which helps refering Event easily.

The event_id just keeps the latest event on the same invocation, meaning that
it refers to the end_test event when the test done, or the start_test event
when test not done or interrupted.

BUG=chrome-os-partner:19335
TEST=Manual
Run Minijack on the build log and check the new event_id field in Test table.

Change-Id: I51f3027d69e64eb31a1e4e6f4442fc2ac52eb85f
Reviewed-on: https://gerrit.chromium.org/gerrit/50776
Reviewed-by: Rong Chang <rongchang@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>