Minijack: Propagate the metadata, containing the log filename, to the end
In the recent refinement of the event log, we can't get the original log
filename from the content of events. We need to propagate the metadata,
containing the log filename to the end. So far, we use the metadata just
for showing the info messages.

BUG=chrome-os-partner:19333
TEST=make test
TEST=Manaul
Run Minijack on the build log and check it produces the same result.

Change-Id: Ie5fc382afe0cc0aa1faae039fcbbac97fe0f6425
Reviewed-on: https://gerrit.chromium.org/gerrit/51091
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>