Minijack: Use image_id and SEQ as event_id
The new log filename on the shopfloor machine is now "events.{image_id}".
The image_id is also an unique id and can also be the event_id. The old log
still works in this approach.

BUG=chrome-os-partner:19333
TEST=Manaul
Run Minijack on the build log and check the event_id changed.

Change-Id: I56c3948cff8dac24df763367968d734bbd106c44
Reviewed-on: https://gerrit.chromium.org/gerrit/51256
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>