Minijack: Base64 encode the event_id
Attr uses event_id as foreign key to Event. As there is a hugh amount of
rows in the Attr table. Reducing the length of event_id can slim the database
size. So use base64 encoding for the event_id field. Also cache the value for
speed-up.

The database size can be reduced ~17% after this change.

BUG=chrome-os-partner:19333
TEST=make test
TEST=Manaul
Run Minijack on the build log and check the event_id changed.

Change-Id: Ia11d33e048088c04a49efb887e66cc2f754c46e1
Reviewed-on: https://gerrit.chromium.org/gerrit/51257
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>