Minijack: Define the __nonzero__ method of EventStream
An EventStream is structure which stores a preamble and a list of
non-preamble events. The class EventStream inherits the list class.
So the default __nonzero__ method checks if the list is non-empty.
It returns False on an empty event lists even if the preamble is valid.

So this change overrides the __nonzero__ method and checks the preamble
in addition.

BUG=chrome-os-partner:19241
TEST=make test

Change-Id: I59f30153016aaedc4768577c21ff02a555ee94af
Reviewed-on: https://gerrit.chromium.org/gerrit/56540
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>