MJFE: Add a set of buttons to select column suites on the build page
On the build page, we have a lot of columns in the device table. It is
not possible to show them all. So we define some column suites, like
"General Info", "Recent Tests", and "Recent Notes". An user can select
the suites using the buttons on top of the table.

This CL also does a bit of style changes and adds the jquery-ui-button
js library which is needed for the buttons.

BUG=chrome-os-partner:19468
TEST=Manaul, manage runserver
Demo: http://go/minijack_frontend_next

Change-Id: I0b7cff547075d828a98cc654819104b86847033c
Reviewed-on: https://gerrit.chromium.org/gerrit/57419
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>