Minijack: Add custom constructors to parse YAML
Instead of using regular expressions to convert the unsupported datatype to
the supported ones, this change adds custom constructors to parse the
unsupported datatypes. They are:
  tuples: treated as lists;
  unicodes/names: treated as strings;
  objects: treated as lists/dicts.

This change also only ignore the unparseable event. The remaining events are
still able to be parsed.

BUG=chrome-os-partner:19898
TEST=make test

Change-Id: I3f98ce7f12b9c47a9c655a8ba52f9cfc14bbc9f4
Reviewed-on: https://gerrit.chromium.org/gerrit/57306
Reviewed-by: Chun-Ta Lin <itspeter@google.com>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>