Minijack: Support indexing on the Minijack data model
We need to create indexes on the tables to speed-up queries. On the Minijack
data model, it needs to support db_index tags to notify fields should be
indexed in the database.

On initializing tables, it checks the indexes exist or not. If not, it
creates them (may take a while if the tables have data).

BUG=chrome-os-partner:19273
TEST=make test

Change-Id: I8a3df3bd099b60335354745715f658e6248d06a3
Reviewed-on: https://gerrit.chromium.org/gerrit/57427
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>