MJFE: Add the detail rows on devices
When an user clicks the icon of a device, it shows the detail rows which
have the controls to ping, ssh, screenshot, and add notes.

This change only makes the UI of the above functions. The implementation
will be done later.

TEST=Manaul, manage runserver
Demo: http://go/minijack_frontend_next

Change-Id: I74c10a3b7c5fba4d5fc307448bccf9c747810959
Reviewed-on: https://gerrit.chromium.org/gerrit/57588
Reviewed-by: Rong Chang <rongchang@chromium.org>
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>